Move common hashtable defitions to another file
We will refactor a hashtable implementation that is used by total_table
and transport_table in ACL. To do this, we first will migrate that code
into a separate file to ease implementation. And for that the common
hashtable methods/defitions will be required.
